methods/dader/requirements.txt
pathtype==1.0.0
torchvision==0.8.2
torchaudio==0.7.2
pandas
demjson==2.2.4
scipy==1.5.4
numpy==1.20.3
transformers==4.6.1
jieba==0.42.1
scikit_learn==0.24.2
